api/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
python-multipart
##########################################################################################
# EXAMPLE (Vision)
pillow
##########################################################################################
